[56] feat : 치아 선택 부분 컴포넌트 분리
