+++ src/components/medical/ToothSelection.module.scss
@@ -13,27 +13,3 @@
   color: $grey3;
   margin-top: 5px;
 }
-.toothItem {
-  margin-top: 7px;
-}
-.toothSelect {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.treatmentButton {
-  border: 1px solid $grey1;
-  border-radius: 8px;
-  padding: 7px 16px;
-  @include font-style($body2);
-  color: $black;
-  cursor: pointer;
-}
-
-.selected {
-  @include button-select;
-}
-.toothImg {
-  margin-top: 12px;
-}
