+++ src/components/medical/ToothSelection.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import Tooth from "../tooth/Tooth";
+import { motion } from "framer-motion";
+import ToothWriteModal from "../modal/ToothWriteModal";
 import styles from "./ToothSelection.module.scss";
 import {
   useCostList,
@@ -9,17 +9,11 @@ import {
   useTreatmentType
 } from "../../stores/medicalWrite";
 import { useModalStore } from "@/stores/modal";
-import ToothWriteModal from "../modal/ToothWriteModal";
 import { ToothType } from "@/constants/toothConstants";
-
-type ToothSide = {
-  value: "left" | "right";
-  name: "왼쪽" | "오른쪽";
-};
+import ToothSelectSelection from "../tooth/ToothSelectSection";
 
 const ToothSelection = () => {
   const { treatmentType } = useTreatmentType();
-  const [toothSelect, setToothSelect] = useState<"left" | "right">("left");
   const [isDisplay, setIsDisplay] = useState<boolean>(false);
   const [isDisplayModal, setIsDisplayModal] = useState<boolean>(false);
   const [selectedTooth, setSelectedTooth] = useState<ToothType>({
@@ -31,6 +25,7 @@ const ToothSelection = () => {
   const { selectedCost } = useCostList();
   const { updateTreatmentList } = useMedicalWriteStore();
   const { treatmentCostList } = useTreatmentCost();
+  const { openModal } = useModalStore();
 
   useEffect(() => {
     const notHasToothId = treatmentCostList
@@ -51,17 +46,6 @@ const ToothSelection = () => {
     updateTreatmentList([...notHasToothId, ...treatmentCost]);
   }, [selectedCost, treatmentCostList]);
 
-  const { openModal } = useModalStore();
-  const toothSide: ToothSide[] = [
-    {
-      value: "left",
-      name: "왼쪽"
-    },
-    {
-      value: "right",
-      name: "오른쪽"
-    }
-  ];
   useEffect(() => {
     const shouldDisplay = !treatmentType.every(
       (treatment) => treatment.name === "스케일링" || treatment.name === "잇몸"
@@ -85,59 +69,24 @@ const ToothSelection = () => {
   return (
     <>
       {isDisplay && (
-        <>
-          <motion.div
-            className={styles.writeWrapper}
-            initial={{ opacity: 0, x: 100 }}
-            animate={{ opacity: 1, x: 0 }}
-            exit={{ opacity: 0, x: -100 }}
-            transition={{ duration: 0.5, ease: "easeInOut" }}
-          >
-            <label className={styles.writeLabel}>
-              치아를 선택해서 받은 치료를 <br /> 설정해 주세요.
-            </label>
-            <span className={styles.helperText}>
-              스케일링과 잇몸치료는 설정에서 제외됩니다.
-            </span>
-            <div className={styles.toothItem}>
-              <div className={styles.toothSelect}>
-                {toothSide.map((side) => (
-                  <button
-                    type="button"
-                    key={side.value}
-                    className={[
-                      styles.treatmentButton,
-                      side.value === toothSelect ? styles.selected : ""
-                    ].join(" ")}
-                    onClick={() => setToothSelect(side.value)}
-                  >
-                    {side.name}
-                  </button>
-                ))}
-              </div>
-              <div className={styles.toothImg}>
-                <AnimatePresence mode="wait">
-                  <motion.div
-                    key={toothSelect}
-                    initial={{ opacity: 0, x: 100 }}
-                    animate={{ opacity: 1, x: 0 }}
-                    exit={{ opacity: 0, x: -100 }}
-                    transition={{ duration: 0.5 }}
-                  >
-                    <Tooth
-                      location={toothSelect}
-                      setIsDisplayModal={setIsDisplayModal}
-                      setSelectedTooth={setSelectedTooth}
-                    />
-                  </motion.div>
-                </AnimatePresence>
-              </div>
-            </div>
-            {/* <span className={styles.errorText}>
-              치아를 눌러 치료를 기록해 주세요.
-            </span> */}
-          </motion.div>
-        </>
+        <motion.div
+          className={styles.writeWrapper}
+          initial={{ opacity: 0, x: 100 }}
+          animate={{ opacity: 1, x: 0 }}
+          exit={{ opacity: 0, x: -100 }}
+          transition={{ duration: 0.5, ease: "easeInOut" }}
+        >
+          <label className={styles.writeLabel}>
+            치아를 선택해서 받은 치료를 <br /> 설정해 주세요.
+          </label>
+          <span className={styles.helperText}>
+            스케일링과 잇몸치료는 설정에서 제외됩니다.
+          </span>
+          <ToothSelectSelection
+            setSelectedTooth={setSelectedTooth}
+            setIsDisplayModal={setIsDisplayModal}
+          />
+        </motion.div>
       )}
     </>
   );
